Fix clippy warnings and improve code quality
- Fix redundant imports and unnecessary casts
- Replace hardcoded PI values with std::f32::consts::PI
- Simplify redundant pattern matching
- Replace vec! with arrays where appropriate
- Fix doc comment syntax in examples
- All tests and benchmarks passing
- Zero clippy warnings remaining

+++ benches/performance_benchmark.rs
@@ -21,7 +21,7 @@ fn bench_batch_read(c: &mut Criterion) {
                 // Mock batch read implementation
                 let mut results = Vec::new();
                 for i in 0..count {
-                    results.push(PlcValue::Dint(i as i32));
+                    results.push(PlcValue::Dint(i));
                 }
                 black_box(results)
             })
